+++ src/main/java/pro/sky/finalprojectsky/service/CommentsService.java
@@ -13,7 +13,13 @@ public class CommentsService {
         this.commentsRepository = commentsRepository;
     }
 
-    public Comment findCommentById(Long id){
+    //Нужен метод для поиска комментария
+    //Проблема в том, что комментарий не отдельная сущность, а часть сущности объявление
+    //И в одном объявлении может быть несколько комментариев
+    //Поэтому, чтобы найти комментарий, надо вначале найти объявление, в котором будем искать комментарий
+    //И уже в найденном объявления искать этот комментарий
+
+    public Comment findCommentById(Integer id){
             return commentsRepository.findById(id).orElse(new Comment());
     }
 }
